Scope Validation errors in support interface to a candidate
To start scoping different validation errors to Providers, Vendor API
and candidates, change the validation errors landing page to a list of
different types of validation errors. And move the current candidate
validation error to a dedicated directory.
